Wrote tests for ProcessTracking find extracts functions
:white_check_mark: Wrote tests for ProcessTracking

Wrote tests for helper functions that are used by ProcessTracking to
find extracts by process name, location name, or full/partial filename.

They are not necessarily part of the core functionality of the ProcessTracking class, but by having them as part of it, users will have an easier time using the framework.