Reconfiguring code to prepare for packaging :construction_worker:  Travis-CI and Coveralls working :building_construction: Moved data_store initialization to DataStore
Verifying that build stack is working, sans package deploy to PyPi.  Initial configuration for PyPi is in travis config, but disabled.
This commit is the first steps for getting the package ready for deploy.
Major refactoring had to be done to ensure that code is only run when
called as opposed to on import.

Closes: #5