Enforced extract record ascending order :bug: Closes: #13 :building_construction: Configured Travis-CI to publish on Github Releases
Resolved #13 where there was a chance where extracts could be returned
in an order other than registration date.  While this was unlikely, the order is now enforced in the queries.

Also tweaked travis-ci to now publish releases to both PyPi and GitHub.  This is based on git tags.  Will not test the publishing until more of the first release backlog is done.