process_tracker_python-1 Add support for other relational databases
:sparkles: MySQL support verified

Tested with MySQL.  Found issue with datetime columns with MySQL where
sub-seconds are not tracked by default.  While SQLAlchemy does have the
ability to take advantage of newer MySQL versions allowing for sub-seconds,
instead of adding complexity to the data model classes the decision was
made to NOT support sub-seconds at this time in MySQL.  If someone else
wants to add it, I'm fine, but hit circular import problem when
determining data store type.

Working on #1