process_tracker_python-30 Extract Lookups should return Extract objects
All the extract file finders now return extract objects so they can be
modified as they are used.  A helper method was also added that provides
the full filename of a given extract (called full_filepath).

Also realized that I missed adding string representations of a bunch of
the data models.