process_tracker_python-50 Extract Finders By Given Status
:bug: Hit connection threshold to database
:sparkles: Extract finders can now search for any status type.

Tweaked ExtractTracker to use the data store and session objects from process_tracker, since an extract tracker must have a process tracker
associated with it.  This stopped the connection threshold problem as
the same connection can be used until the process tracker changes to 'completed' or 'failed' status, which is when the session is closed.

Finder methods now have a default status of 'ready' (which should be the
common status required) but any status type can be searched for, which
allows custom status types to be leveraged for extracts.

Closes #50