process_tracker_python-63 Funky Calls On process_tracker Import
Cleaned up errors.  Part of problem was adding passive_deletes in another ticket. Cleaned up the ones that were causing a warning.  Need to dive deeper as to what needs to be done for those.

Need to figure out how to get around the import loop between aws_utilities and settings.  For now, hard set the aws_utilities log_level to INFO.

Closes #63